% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_decomp_cond.R
\name{bs_decomp_cond}
\alias{bs_decomp_cond}
\title{conditional Brier score decomposition}
\usage{
bs_decomp_cond(p, o, states, bins = NULL, method = "bias_corrected")
}
\arguments{
\item{p}{vector of probability forecasts.}

\item{o}{vector of binary outcomes.}

\item{states}{vector of states on which to perform the conditional decomposition.}

\item{bins}{number of bins to use in the decomposition.}

\item{method}{string specifying which method should be used to estimate the decomposition terms.}
}
\value{
A named vector containing the decomposition terms.
}
\description{
conditional Brier score decomposition
}
\examples{
p <- runif(100)
o <- rbinom(100, 1, 0.5)
p_disc <- sample(seq(0, 1, 0.1), 100, TRUE)
states = rep(c("G1", "G2"), each = 50)

bs_decomp_cond(p = p, o = o, states = states, bins = 10)
bs_decomp_cond(p = p_disc, o = o, states = states)
bs_decomp_cond(p = p, o = o, states = states, bins = 10, method = "classical")
bs_decomp_cond(p = p, o = o, states = states, method = "isotonic")
}
