% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{waterfall plot}
\usage{
waterfall_plot(
  p,
  o,
  states,
  bins = NULL,
  title = "",
  scale = 10000,
  alt = FALSE
)
}
\arguments{
\item{p}{vector of probability forecasts.}

\item{o}{vector of binary outcomes.}

\item{states}{vector of states on which to perform the conditional decomposition.}

\item{bins}{number of bins to use in the decomposition.}

\item{title}{optional title of the plot.}

\item{scale}{amount by which to scale the decomposition terms.}

\item{alt}{logical specifying which of the two waterfall plots should be used.}
}
\description{
to show Brier decomposition terms
}
\examples{
p <- runif(100)
o <- rbinom(100, 1, 0.5)
p_disc <- sample(seq(0, 1, 0.1), 100, TRUE)
states <- rep(c("G1", "G2"), each = 50)

waterfall_plot(p = p_disc, o = o, states = states)
waterfall_plot(p = p_disc, o = o, states = states, alt = TRUE)
}
