% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_decomp.R
\name{bs_decomp}
\alias{bs_decomp}
\title{unconditional Brier score decomposition terms}
\usage{
bs_decomp(p, o, bins = NULL, method = "bias_corrected")
}
\arguments{
\item{p}{vector of probability forecasts.}

\item{o}{vector of binary outcomes.}

\item{bins}{number of bins to use in the decomposition.}

\item{method}{string specifying which method should be used to estimate the decomposition terms.}
}
\value{
A named vector containing the decomposition terms.
}
\description{
unconditional Brier score decomposition terms
}
\examples{
p <- runif(100)
o <- rbinom(100, 1, 0.5)
p_disc <- sample(seq(0, 1, 0.1), 100, TRUE)

bs_decomp(p = p, o = o, bins = 10)
bs_decomp(p = p_disc, o = o)
bs_decomp(p = p, o = o, bins = 10, method = "classical")
bs_decomp(p = p, o = o, method = "isotonic")
}
