% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_decomp.R
\name{bs_decomp}
\alias{bs_decomp}
\alias{bs_decomp_cond}
\title{Decompositions of the Brier score}
\usage{
bs_decomp(o, p, bins = NULL, method = "isotonic")

bs_decomp_cond(o, p, states, bins = NULL, method = "bias-corrected")
}
\arguments{
\item{o}{vector of binary outcomes.}

\item{p}{vector of probability forecasts.}

\item{bins}{integer; number of bins to use in the decomposition.}

\item{method}{string; the method to be used to estimate the decomposition terms.}

\item{states}{vector of states on which to perform the conditional decomposition.}
}
\value{
A named vector containing the decomposition terms.
}
\description{
Calculate decompositions of the Brier score into uncertainty, resolution, and
reliability components. A conditional decomposition can also be performed that
calculates these terms conditional on some events, or states having occurred.
}
\details{
The vectors \code{p}, \code{o}, and \code{states} (if used) should have the
same length.

\code{o} is a numeric vector of values that are either 0 or 1, representing
the observed values.
\code{p} is a numeric vector of values between 0 and 1, representing the
corresponding probability forecasts.
\code{states} is a character vector, with \code{states[i]} corresponding to
the state that occurs when \code{p[i]} is forecast and \code{o[i]} is observed.
The number of unique states should be small compared to the number of observations.

The \code{bins} argument specifies how many bins will be used when calculating
the decomposition. The default is the number of unique elements in \code{p}.
A warning is returned if this is large compared to the number of observations.
This argument is not required when the decomposition is calculated using
isotonic regression (\code{method = "isotonic"}).

The \code{method} argument takes three possible options: \code{method = "classical"},
performs the classical decomposition of the Brier score that was proposed by
Murphy (1973); \code{method = "bias-corrected"} performs the bias-corrected
decomposition proposed by Ferro and Fricker (2012), which is more appropriate
for small sample sizes; \code{method = "isotonic"} (the default) performs the
decomposition based on isotonic regression proposed by Dimitriadis et al. (2021).
See references below. Note that the bias-corrected approach may return negative
estimates of the three terms.

Calculation of the Brier score decompositions using isotonic regression
leverages the \pkg{reliabilitydiag} package.
}
\examples{
p <- runif(100)
o <- rbinom(100, 1, 0.5)
p_disc <- sample(seq(0, 1, 0.1), 100, TRUE)

# unconditional decomposition

bs_decomp(o = o, p = p, bins = 10)
bs_decomp(o = o, p = p_disc)
bs_decomp(o = o, p = p, bins = 10, method = "classical")
bs_decomp(o = o, p = p, method = "isotonic")

# conditional decomposition

states = rep(c("G1", "G2"), each = 50)

bs_decomp_cond(o = o, p = p, states = states, bins = 10)
bs_decomp_cond(o = o, p = p_disc, states = states)
bs_decomp_cond(o = o, p = p, states = states, bins = 10, method = "classical")
bs_decomp_cond(o = o, p = p, states = states, method = "isotonic")

}
\references{
\emph{Decomposition of the Brier score}

Murphy, A.H. (1973): `A new vector partition of the probability score'. \emph{Journal of Applied Meteorology and Climatology} 12, 595-600. \doi{10.1175/1520-0450(1973)012<0595:ANVPOT>2.0.CO;2}

\emph{Bias-corrected decomposition of the Brier score}

Ferro, C.A.T. and T.E. Fricker (2012): `A bias‐corrected decomposition of the Brier score. \emph{Quarterly Journal of the Royal Meteorological Society} 138, 1954-1960. \doi{10.1002/qj.1924}

\emph{Isotonic regression-based decomposition of the Brier score}

Dimitriadis, T., Gneiting, T. and A.I. Jordan (2021): `Stable reliability diagrams for probabilistic classifiers'. \emph{Proceedings of the National Academy of Sciences} 118, e2016191118. \doi{10.1073/pnas.2016191118}

\emph{Decomposition of proper scoring rules}

Bröcker, J. (2009): `Reliability, sufficiency, and the decomposition of proper scores'. \emph{Quarterly Journal of the Royal Meteorological Society} 135, 1512-1519. \doi{10.1002/qj.456}

\emph{Conditional decomposition of proper scoring rules}

Allen, S., Ferro, C.A.T. and F. Kwasniok (2023): `A conditional decomposition of proper scores: quantifying the sources of information in a forecast'. \emph{Quarterly Journal of the Royal Meteorological Society}
}
\author{
Sam Allen
}
